prevents creation of resource descriptors for events tables and removes existing schemas
the combinatoriality of column names is very high and a few columns would require a case distinction because, for example, the events table my include negative 'duration' values (for <location> tags) etc. Since events tables simply encode the entire XML structure, it does not make a lot of sense to define schemas for each possible combination of columns, which really blew up the amount of schemas required in the previous attempt. On top of that, most fields did not have a description and simply defaulted to type 'string'